small refactor
- change structure
- add auto format on save
- add lazygit plugin
