+++ init.lua
@@ -1,1 +1,74 @@
-require("theprimeagen")
+require("set")
+require("remap")
+require("lazy_init")
+require('mini.statusline').setup()
+
+local augroup = vim.api.nvim_create_augroup
+local TheGroup = augroup('TheGroup', {})
+
+local autocmd = vim.api.nvim_create_autocmd
+local yank_group = augroup('HighlightYank', {})
+
+function R(name)
+    require("plenary.reload").reload_module(name)
+end
+
+autocmd('TextYankPost', {
+    group = yank_group,
+    pattern = '*',
+    callback = function()
+        vim.highlight.on_yank({
+            higroup = 'IncSearch',
+            timeout = 40,
+        })
+    end,
+})
+
+-- Remove trailing whitespace on save
+autocmd({ "BufWritePre" }, {
+    group = TheGroup,
+    pattern = "*",
+    command = [[%s/\s\+$//e]],
+})
+
+-- Format on save
+autocmd({ "BufWritePre" }, {
+    group = TheGroup,
+    pattern = "*",
+    callback = function()
+        vim.lsp.buf.format { async = false }
+    end
+})
+
+autocmd('BufEnter', {
+    group = TheGroup,
+    callback = function()
+        if vim.bo.filetype == "zig" then
+            vim.cmd.colorscheme("tokyonight-night")
+        else
+            vim.cmd.colorscheme("rose-pine-moon")
+        end
+    end
+})
+
+autocmd('LspAttach', {
+    group = TheGroup,
+    callback = function(e)
+        local opts = { buffer = e.buf }
+        vim.keymap.set("n", "gd", function() vim.lsp.buf.definition() end, opts)
+        vim.keymap.set("n", "K", function() vim.lsp.buf.hover() end, opts)
+        vim.keymap.set("n", "<leader>vws", function() vim.lsp.buf.workspace_symbol() end, opts)
+        vim.keymap.set("n", "<leader>vd", function() vim.diagnostic.open_float() end, opts)
+        vim.keymap.set("n", "<leader>vca", function() vim.lsp.buf.code_action() end, opts)
+        vim.keymap.set("n", "<leader>vfr", function() vim.lsp.buf.references() end, opts)
+        vim.keymap.set("n", "<leader>vrn", function() vim.lsp.buf.rename() end, opts)
+        vim.keymap.set("i", "<C-h>", function() vim.lsp.buf.signature_help() end, opts)
+        vim.keymap.set("n", "[d", function() vim.diagnostic.goto_next() end, opts)
+        vim.keymap.set("n", "]d", function() vim.diagnostic.goto_prev() end, opts)
+    end
+})
+
+vim.g.netrw_browse_split = 0
+vim.g.netrw_banner = 0
+vim.g.netrw_winsize = 25
+vim.o.splitright = true
