+++ lazy_init.lua
@@ -12,6 +12,9 @@ end
 vim.opt.rtp:prepend(lazypath)
 
 require("lazy").setup({
-    spec = "theprimeagen.lazy",
+    spec = {
+        -- import your plugins
+        { import = "plugins" },
+    },
     change_detection = { notify = false }
 })
