+++ lua/plugins/cinnamon.lua
