+++ lua/plugins/colors.lua
