+++ lua/plugins/ministatusline.lua
