+++ lua/plugins/trouble.lua
